+++ src/manager.ts
@@ -9,7 +9,7 @@ export class ExtensionManager implements vscode.Disposable {
   private _context: vscode.ExtensionContext;
   private _channel: vscode.OutputChannel;
   public logger: Logger = new Logger();
-  private _terminal: vscode.Terminal | null = null;
+  private _isRaycastEnabled = false;
 
   constructor(public readonly extensionContext: vscode.ExtensionContext) {
     this._context = extensionContext;
@@ -24,6 +24,7 @@ export class ExtensionManager implements vscode.Disposable {
       })
     );
     this.registerPackageJsonChanges();
+    this.registerCompletionProviders();
   }
 
   private registerPackageJsonChanges() {
@@ -85,6 +86,7 @@ export class ExtensionManager implements vscode.Disposable {
         }
       }
     }
+    this._isRaycastEnabled = isRaycastEnabled;
     this.logger.debug(`${pkgjsonFilename} raycast enabled: ${isRaycastEnabled}`);
     await this.setContext("raycast.workspaceEnabled", isRaycastEnabled);
   }
@@ -159,6 +161,10 @@ export class ExtensionManager implements vscode.Disposable {
     return result;
   }
 
+  get isRaycastEnabled(): boolean {
+    return this._isRaycastEnabled;
+  }
+
   public runNpmExec(cmd: string[], terminalID?: string | undefined) {
     const term = this.getTerminal(terminalID);
     term.sendText("clear");
@@ -192,11 +198,79 @@ export class ExtensionManager implements vscode.Disposable {
 
   public async getImageAssets(): Promise<string[]> {
     const ws = this.getActiveWorkspace();
-    if (!ws) {
+    if (!ws || !this.isRaycastEnabled) {
       return [];
     }
     const assetsFolder = path.join(ws.uri.fsPath, "assets");
     return await getImageAssetsFromFolder(assetsFolder);
+  }
+
+  private registerCompletionProviders() {
+    const self = this;
+    const tsImageAssetCompletionProvider = vscode.languages.registerCompletionItemProvider(
+      "typescriptreact",
+      {
+        async provideCompletionItems(
+          document: vscode.TextDocument,
+          position: vscode.Position,
+          token: vscode.CancellationToken,
+          context: vscode.CompletionContext
+        ) {
+          const line = document.lineAt(position.line);
+          const text = line.text.substring(0, position.character);
+          const sourceIndex = text.lastIndexOf("source:");
+          const iconIndex = text.lastIndexOf("icon=");
+          if ((sourceIndex > 0 && text.length - sourceIndex < 15) || (iconIndex > 0 && text.length - iconIndex < 15)) {
+            const assets = await self.getImageAssets();
+            if (assets && assets.length > 0) {
+              return assets.map((a) => new vscode.CompletionItem(a, vscode.CompletionItemKind.File));
+            }
+          }
+          return undefined;
+        },
+      },
+      '"',
+      "'"
+    );
+    const jsonImageAssetCompletionProvider = vscode.languages.registerCompletionItemProvider(
+      "json",
+      {
+        async provideCompletionItems(
+          document: vscode.TextDocument,
+          position: vscode.Position,
+          token: vscode.CancellationToken,
+          context: vscode.CompletionContext
+        ) {
+          const filename = path.basename(document.fileName);
+          if (filename !== "package.json") {
+            return undefined;
+          }
+          const line = document.lineAt(position.line);
+          const text = line.text.substring(0, position.character);
+          const lastColon = text.lastIndexOf(":");
+          if (lastColon > 0) {
+            const splits = text.substring(0, lastColon).split(":");
+            if (splits && splits.length > 0) {
+              const last = splits[splits.length - 1];
+              if (last.includes('"icon"')) {
+                const assets = await self.getImageAssets();
+                if (assets && assets.length > 0) {
+                  const comps = assets.map((a) => {
+                    const c = new vscode.CompletionItem(a, vscode.CompletionItemKind.File);
+                    c.range = new vscode.Range(position, position); // required for json files
+                    return c;
+                  });
+                  return comps;
+                }
+              }
+            }
+          }
+          return undefined;
+        },
+      },
+      '"'
+    );
+    this._context.subscriptions.push(tsImageAssetCompletionProvider, jsonImageAssetCompletionProvider);
   }
 
   public dispose() {}